Fix Division by Zero in Secession Event Progress
I've encountered an issue in my current save where the game crashes after I lose a market in the faction secession event. The crash occurs on the next `HoldoutEventFactor.getProgress` call after losing the market.

The issue is that when the ratio of marketsStillOwned to getOriginalMarkets.size is < 1.0, the int cast will produce 0, at which point we divide `BASE_POINTS` by zero.

This patch has fixed the issue in my save.